requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
jinja2==3.1.2
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.25.2
python-multipart==0.0.6
huggingface_hub==0.14.1
transformers==4.30.2
python-dotenv==1.0.0
pydantic-settings>=2.0.0
httpx>=0.24.1
black==23.7.0
isort==5.12.0
ftfy==6.1.3
regex==2023.8.8
torch==2.0.1
Pillow==10.0.0
clip @ git+https://github.com/openai/CLIP.git
pytest>=7.4.0
pytest-asyncio>=0.21.1
